% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bspline.R
\name{Bs}
\alias{Bs}
\title{Generate a time series of fractional Brownian motion.}
\usage{
Bs(n, m)
}
\arguments{
\item{n}{the number of data which are simulated by f=sin(12*(x+0.2))/(x+0.2)}

\item{m}{the hurst index, with the default value 0.71}
}
\value{
numeric vector
}
\description{
This function generatea a time series of one dimension fractional Brownian motion.
adapted from http://www.mathworks.com.au/matlabcentral/fileexchange/38935-fractional-brownian-motion-generator .
}
\examples{
Bs(100,6)
}
