% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brown.R
\name{fbm}
\alias{fbm}
\title{Generate a time series of fractional Brownian motion.}
\usage{
fbm(hurst = 0.7, n = 100)
}
\arguments{
\item{hurst}{the hurst index, with the default value 0.71}

\item{n}{the number of points between 0 and 1 that will be generated, with the default value 100}
}
\value{
numeric vector
}
\description{
This function generatea a time series of one dimension fractional Brownian motion.
}
\examples{
d <- fbm(hurst=0.5, n=500)
plot(d)
}
