% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ha-package.R
\docType{package}
\name{ha-package}
\alias{ha}
\alias{ha-package}
\title{ha: Interact with 'Home Assistant'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides tools to control smart home features via the API for 'Home Assistant' \url{https://developers.home-assistant.io/}. Allows for controlling states, services, and events. Supports configuring both local access and remote access.
}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
