% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ha_api}
\alias{ha_api}
\title{Check if the API is running}
\usage{
ha_api(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the API information.
}
\description{
Check if the API is running
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_api()
\dontshow{\}) # examplesIf}
}
