% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calendar.R
\name{ha_api_calendars}
\alias{ha_api_calendars}
\title{Retrieve Home Assistant calendar}
\usage{
ha_api_calendars(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant calendar.
}
\description{
Retrieve Home Assistant calendar
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_api_calendars()
\dontshow{\}) # examplesIf}
}
