% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_config.R
\name{ha_api_config}
\alias{ha_api_config}
\title{Retrieve Home Assistant Configuration}
\usage{
ha_api_config(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant configuration.
}
\description{
Retrieve Home Assistant Configuration
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_api_config()
\dontshow{\}) # examplesIf}
}
