% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_states.R
\name{ha_api_states}
\alias{ha_api_states}
\title{Retrieve Home Assistant States}
\usage{
ha_api_states(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant configuration.
}
\description{
Retrieve Home Assistant States
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_api_states()
\dontshow{\}) # examplesIf}
}
