% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{ha_check_config}
\alias{ha_check_config}
\title{Get Home Assistant configuration validity}
\usage{
ha_check_config(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant configuration check result.
}
\description{
Get Home Assistant configuration validity
}
\examples{
\dontshow{if (ha_has_key()) withAutoprint(\{ # examplesIf}
ha_check_config()
\dontshow{\}) # examplesIf}
}
