% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_states.R
\name{ha_delete_states}
\alias{ha_delete_states}
\title{Remove Home Assistant State}
\usage{
ha_delete_states(entity_id, ip = NULL)
}
\arguments{
\item{entity_id}{Entity ID. For example, \code{sensor.temperature}.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing a confirmation message.
}
\description{
Remove Home Assistant State
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_set_states('rpkg.test', 'on')
ha_delete_states('rpkg.test')
\dontshow{\}) # examplesIf}
}
