% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calendar.R
\name{ha_get_calendars}
\alias{ha_get_calendars}
\title{Retrieve Home Assistant calendars}
\usage{
ha_get_calendars(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing Home Assistant calendar information.
}
\description{
Retrieve Home Assistant calendars
}
\examples{
\dontshow{if (ha_has_key()) withAutoprint(\{ # examplesIf}
ha_get_calendars()
\dontshow{\}) # examplesIf}
}
