% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{ha_get_config}
\alias{ha_get_config}
\title{Retrieve Home Assistant Configuration}
\usage{
ha_get_config(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant configuration.
}
\description{
Retrieve Home Assistant Configuration
}
\examples{
\dontshow{if (ha_has_key()) withAutoprint(\{ # examplesIf}
ha_get_config()
\dontshow{\}) # examplesIf}
}
