% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_log.R
\name{ha_get_error_log}
\alias{ha_get_error_log}
\title{Retrieve Home Assistant Error Log}
\usage{
ha_get_error_log(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant error log
}
\description{
Retrieve Home Assistant Error Log
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_get_error_log()
\dontshow{\}) # examplesIf}
}
