% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{ha_get_events}
\alias{ha_get_events}
\title{Retrieve Home Assistant Events}
\usage{
ha_get_events(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing Home Assistant event information.
}
\description{
Retrieve Home Assistant Events
}
\examples{
\dontshow{if (ha_has_key()) withAutoprint(\{ # examplesIf}
ha_get_events()
\dontshow{\}) # examplesIf}
}
