% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_history.R
\name{ha_get_history}
\alias{ha_get_history}
\title{Retrieve Home Assistant History}
\usage{
ha_get_history(entity_id, timestamp = NULL, ip = NULL)
}
\arguments{
\item{entity_id}{Required. Entity ID. For example, \code{sensor.temperature}.}

\item{timestamp}{Optional. Time in \code{YYYY-MM-DDThh:mm:ssTZD} format.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant configuration.
}
\description{
Retrieve Home Assistant History
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_get_history('update.home_assistant_core_update')
\dontshow{\}) # examplesIf}
}
