% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logbook.R
\name{ha_get_logbook}
\alias{ha_get_logbook}
\title{Retrieve Home Assistant Logbook Entries}
\usage{
ha_get_logbook(timestamp = NULL, ip = NULL)
}
\arguments{
\item{timestamp}{Optional. Time in \code{YYYY-MM-DDThh:mm:ssTZD} format.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing Home Assistant logbook entries.
}
\description{
Retrieve Home Assistant Logbook Entries
}
\examples{
\dontshow{if (ha_has_key()) withAutoprint(\{ # examplesIf}
ha_get_logbook()
\dontshow{\}) # examplesIf}
}
