% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{ha_render_template}
\alias{ha_render_template}
\title{Render a Home Assistant Template}
\usage{
ha_render_template(template, ip = NULL)
}
\arguments{
\item{template}{a length one character vector containing the template to render.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A length one character vector containing the rendered template
}
\description{
Render a Home Assistant Template
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_render_template("My living room light is {{ states('light.h6076') }}.")
\dontshow{\}) # examplesIf}
}
