% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_event.R
\name{ha_set_event}
\alias{ha_set_event}
\title{Set Home Assistant Services}
\usage{
ha_set_event(event_type, event_data = NULL, ip = NULL)
}
\arguments{
\item{event_type}{The type of event to set, e.g. \code{light.turn_on}}

\item{event_data}{Optional. A named list of data to send with the event.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A vector of events fired with a confirmation or detailed information
}
\description{
Set Home Assistant Services
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_set_event('light.test')
\dontshow{\}) # examplesIf}
}
