% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_intent.R
\name{ha_set_intent}
\alias{ha_set_intent}
\title{Set Home Assistant Intents}
\usage{
ha_set_intent(intent, ip = NULL)
}
\arguments{
\item{intent}{an intent to fire}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list of intents
}
\description{
Note that this requires use of "intents" for Home Assistant, which
are not enabled by default. Using these without enabling them will
create an error.
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_set_intent(list(name = 'SetTimer', data = list(seconds = '30')))
\dontshow{\}) # examplesIf}
}
