% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_services.R
\name{ha_set_services}
\alias{ha_set_services}
\title{Set Home Assistant Services}
\usage{
ha_set_services(domain, service, entity_id, ip = NULL)
}
\arguments{
\item{domain}{An entity domain. For example, \code{light} or \code{switch}.}

\item{service}{A service name. For example, \code{turn_on} or \code{turn_off}.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list of changes made by the service call
}
\description{
Set Home Assistant Services
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_set_services('light.h6076'
\dontshow{\}) # examplesIf}
}
