% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_states.R
\name{ha_set_states}
\alias{ha_set_states}
\title{Set Home Assistant State}
\usage{
ha_set_states(entity_id, state, attributes = NULL, ip = NULL)
}
\arguments{
\item{entity_id}{Entity ID. For example, \code{sensor.temperature}.}

\item{state}{A state value. For example, \code{on} or \code{off}.}

\item{attributes}{Optional. A named list of attributes to set for the entity.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
a list of information on the new or updated state
}
\description{
Set Home Assistant State
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_set_states(entity_id = 'rpkg.test', state = 'on',
  attributes = list(version = '4.5.1'))
\dontshow{\}) # examplesIf}
}
