% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{key}
\alias{key}
\alias{ha_has_key}
\alias{ha_get_key}
\title{Check or Get Home Assistant API Key}
\usage{
ha_has_key()

ha_get_key()
}
\value{
logical if \code{has}, key if \code{get}
}
\description{
Check or Get Home Assistant API Key
}
\examples{
ha_has_key()
}
\concept{key}
