% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bspline.R
\name{Bs}
\alias{Bs}
\title{Compare different orders k in B-spline regression functions with any m equidistant inner knots}
\usage{
Bs(n, m)
}
\arguments{
\item{n}{the number of data which are simulated by f=sin(12*(x+0.2))/(x+0.2)}

\item{m}{regression functions with m equidistant inner knots}
}
\value{
numeric vector
}
\description{
This function is B-spline regression functions with any m equidistant inner knots and different orders k = 1 to 6.
}
\examples{
Bs(100,6)
}
