% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{ha_get_events}
\alias{ha_get_events}
\title{Retrieve Home Assistant Events}
\usage{
ha_get_events(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant events.
}
\description{
Retrieve Home Assistant Events
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_get_events()
\dontshow{\}) # examplesIf}
}
