% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_services.R
\name{ha_get_services}
\alias{ha_get_services}
\title{Retrieve Home Assistant Services}
\usage{
ha_get_services(ip = NULL)
}
\arguments{
\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant configuration.
}
\description{
Retrieve Home Assistant Services
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_get_services()
\dontshow{\}) # examplesIf}
}
