% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_states.R
\name{ha_get_states}
\alias{ha_get_states}
\title{Retrieve Home Assistant States}
\usage{
ha_get_states(entity_id = NULL, ip = NULL)
}
\arguments{
\item{entity_id}{Optional. Entity ID. For example, \code{sensor.temperature}.}

\item{ip}{IP address. Default is \code{NULL}, which uses the local host.}
}
\value{
A list containing the Home Assistant configuration.
}
\description{
Retrieve Home Assistant States
}
\examples{
\dontshow{if (ha_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ha_get_states()
\dontshow{\}) # examplesIf}
}
